refactor: split chat and iframe logic
